feature(pauseScreen): putting structs in individual files
